requirements.txt
matplotlib>=3.10.3
PyYAML>=6.0.2